refactor scf to put dm training into addons
